requirements.txt
transformers
google.generativeai
streamlit
os
dotenv
openai
PyPDF2
docx
re
requests
bs4